Improve include/poac/io/config.hpp

+++ include/poac/opts/publish.hpp
@@ -39,7 +39,7 @@ namespace poac::opts::publish {
         std::string repo;
         semver::Version version;
         std::optional<std::string> description;
-        std::uint16_t cpp_version;
+        std::uint8_t cpp_version;
         std::optional<std::string> license;
         io::lockfile::PackageType package_type;
         std::string local_commit_sha;
@@ -154,7 +154,7 @@ namespace poac::opts::publish {
     }
 
     [[nodiscard]] std::optional<core::except::Error>
-    verify_cpp_version(const std::uint16_t& cpp_version) {
+    verify_cpp_version(const std::uint8_t& cpp_version) {
         switch (cpp_version) {
             case 98:
                 [[fallthrough]];
@@ -262,7 +262,7 @@ namespace poac::opts::publish {
         }
     }
 
-    std::uint16_t
+    std::uint8_t
     get_cpp_version(const std::optional<io::config::Config>& config) {
         return config->cpp_version.value();
     }
